Revert "Merge pr-4998"
This reverts commit bef61169cefc5ba63598df50ab02990d6a340c06, reversing
changes made to a5a97f586801c861adc19d92a2410a798beb1982.
