+++ tools/plugins/com.liferay.ide.project.core/src/com/liferay/ide/project/core/service/TargetPlatformPossibleValuesService.java
@@ -12,11 +12,12 @@
  * details.
  */
 
-package com.liferay.ide.project.core.workspace;
+package com.liferay.ide.project.core.service;
 
 import com.liferay.ide.core.util.SapphireContentAccessor;
 import com.liferay.ide.core.util.SapphireUtil;
 import com.liferay.ide.core.workspace.WorkspaceConstants;
+import com.liferay.ide.project.core.workspace.NewLiferayWorkspaceOp;
 
 import java.util.ArrayList;
 import java.util.Collections;
